+++ CHANGELOG.md
@@ -3,7 +3,28 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Upcoming
+## v0.4
+
+### Added
+- Added grouped bar charts
+
+## v0.3
+
+### Added
+- Option for automatic tick marks based on chart size
+- Added Tooltips API
+- Added default Tooltip
+- Added Event.off
+
+### Changed
+- X-Axis labels rotated counter-clockwise (to use Y-Axis space)
+- Changed 'autolabels' default to true
+- Changed 'renderer' default to 'svg'
+
+### Support
+- Added example demo links to JSFiddle
+
+## v0.2
 
 ### Added
 - Sparkline chart type